+++ anom/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-from . import properties, properties as props
+from . import conditions, properties, properties as props
 from .adapter import Adapter, get_adapter, set_adapter
 from .model import Key, Model, Property, delete_multi, get_multi, put_multi, lookup_model_by_kind
 from .query import Query, Resultset, Page, Pages